Added factorials.java
This is a java program to find the factorial of really large numbers for which factorial can't be calculated by normal operations. This program can also calculate the factorials up to the mark of 10^18 or in words it is incredibly 1 billion billion.

The idea behind the working of this program is that using the traditional method for multiplying numbers, the same method we used in primary classes, instead of recursion (the modern way). And the number is so large that it has been stored in array of large length in the form of separate digits. Length of array depends on input. Sufficient length would be i*n(i) where i is the number whose factorial we want to find and n(i) denotes the no. of digits in i

But such a factorial is not practical because it may take years to calculate so I've just limited it to integer values that is around 1 billion. The time taken in calculation such factorials increases exponentially that is clear from the video (link given below). In the video I've skip